Rename chroma effect and chroma SDK to be prefixed with Old- and moved to "old" folder
