+++ apps/website/src/app/services/chromaEffect/state/multi-dynamic-state/multi-dynamic-state.ts
@@ -1,6 +1,6 @@
 import { State } from '../abstract/state';
 import { color_wheel, randomInteger } from '../../../../shared/functions';
-import { HeadsetEffect, KeyboardEffect, MouseEffect } from '../../../chromaSDK/chromaSDK.service';
+import { HeadsetEffect, KeyboardEffect, MouseEffect } from '../../../old/chromaSDK/old-chroma-s-d-k.service';
 
 export class MultiDynamicState extends State {
   private headsetColors = [0, 0, 0, 0, 0];
