+++ apps/website/src/app/services/chromaEffect/state/single-dynamic-state/single-dynamic-state.ts
@@ -1,6 +1,6 @@
 import { State } from '../abstract/state';
 import { color_wheel, randomInteger } from '../../../../shared/functions';
-import { HeadsetEffect, KeyboardEffect, MouseEffect } from '../../../chromaSDK/chromaSDK.service';
+import { HeadsetEffect, KeyboardEffect, MouseEffect } from '../../../old/chromaSDK/old-chroma-s-d-k.service';
 
 export class SingleDynamicState extends State {
   private interval: string | number | NodeJS.Timeout | undefined;
