+++ apps/website/src/app/services/old/chromaEffect/old-chroma-effect.service.ts
@@ -1,18 +1,21 @@
 import { Injectable } from '@angular/core';
-import { ChromaSDKService } from '../chromaSDK/chromaSDK.service';
-import { State } from './state/abstract/state';
-import { StaticState } from './state/static-state/static-state';
+import { OldChromaSDKService } from '../chromaSDK/old-chroma-s-d-k.service';
+import { State } from '../../chromaEffect/state/abstract/state';
+import { StaticState } from '../../chromaEffect/state/static-state/static-state';
 import iro from '@jaames/iro';
-import { VisualizerState } from './state/visualizer-state/visualizer-state';
-import { MessageService } from '../message-service/message.service';
-import { VisualizerBrightnessState } from './state/visualizer-brightness-state/visualizer-brightness-state';
+import { VisualizerState } from '../../chromaEffect/state/visualizer-state/visualizer-state';
+import { MessageService } from '../../message-service/message.service';
+import { VisualizerBrightnessState } from '../../chromaEffect/state/visualizer-brightness-state/visualizer-brightness-state';
 import { Store } from '@ngrx/store';
-import { UserPreferences } from '../../shared/types/types';
+import { UserPreferences } from '../../../shared/types/types';
 
+/**
+ * @deprecated
+ */
 @Injectable({
   providedIn: 'root'
 })
-export class ChromaEffectService extends ChromaSDKService {
+export class OldChromaEffectService extends OldChromaSDKService {
   constructor(messageService: MessageService, store: Store<{ userPreferences: UserPreferences }>) {
     super(messageService, store);
     this._state = new StaticState();
